Refactoring: remove tribler_core/session.py to simplify code structure
